+++ src/__tests__/button-test.tsx
@@ -1,10 +1,12 @@
 import * as React from 'react';
 import {ThemeVariant} from '../theme-variant-context';
-import {ButtonPrimary, ButtonSecondary} from '../button';
+import {ButtonDanger, ButtonLink, ButtonPrimary, ButtonSecondary} from '../button';
 import * as Spinner from '../spinner';
 import {render, screen} from '@testing-library/react';
 import ThemeContextProvider from '../theme-context-provider';
 import {makeTheme} from './test-utils';
+import userEvent from '@testing-library/user-event';
+import IconPhotoCameraRegular from '../generated/mistica-icons/icon-photo-camera-regular';
 
 test('button is accesible', () => {
     render(
@@ -405,4 +407,122 @@ test('dark theme', () => {
           </a>
         </DocumentFragment>
     `);
+});
+
+test('buttons can track events', async () => {
+    const logEventSpy = jest.fn();
+    const trackingEvent = {
+        category: 'someCategory',
+        action: 'someAction',
+        label: 'someLabel',
+    };
+
+    render(
+        <ThemeContextProvider theme={makeTheme({analytics: {logEvent: logEventSpy}})}>
+            <ButtonPrimary trackingEvent={trackingEvent} href="#test">
+                button with href
+            </ButtonPrimary>
+            <ButtonPrimary trackingEvent={trackingEvent} to="#test">
+                button with to
+            </ButtonPrimary>
+            <ButtonPrimary trackingEvent={trackingEvent} onPress={() => {}}>
+                button with onPress
+            </ButtonPrimary>
+        </ThemeContextProvider>
+    );
+
+    const hrefButton = await screen.findByRole('button', {name: 'button with href'});
+    const toButton = await screen.findByRole('button', {name: 'button with to'});
+    const onPressButton = await screen.findByRole('button', {name: 'button with onPress'});
+
+    userEvent.click(hrefButton);
+    expect(logEventSpy).toHaveBeenCalledWith(trackingEvent);
+    expect(logEventSpy).toHaveBeenCalledTimes(1);
+
+    userEvent.click(toButton);
+    expect(logEventSpy).toHaveBeenCalledWith(trackingEvent);
+    expect(logEventSpy).toHaveBeenCalledTimes(2);
+
+    userEvent.click(onPressButton);
+    expect(logEventSpy).toHaveBeenCalledWith(trackingEvent);
+    expect(logEventSpy).toHaveBeenCalledTimes(3);
+});
+
+test('buttons track default events', async () => {
+    const logEventSpy = jest.fn();
+
+    render(
+        <ThemeContextProvider theme={makeTheme({analytics: {logEvent: logEventSpy}})}>
+            <ButtonPrimary href="#test">no track</ButtonPrimary>
+            <ButtonPrimary trackEvent href="#test">
+                primary
+            </ButtonPrimary>
+            <ButtonSecondary trackEvent to="#test">
+                secondary
+            </ButtonSecondary>
+            <ButtonDanger trackEvent onPress={() => {}}>
+                danger
+            </ButtonDanger>
+            <ButtonPrimary trackEvent onPress={() => {}}>
+                <IconPhotoCameraRegular color="currentColor" />
+                Take a photo
+            </ButtonPrimary>
+            <ButtonLink href="#">no track link</ButtonLink>
+            <ButtonLink trackEvent href="#">
+                link
+            </ButtonLink>
+        </ThemeContextProvider>
+    );
+
+    const noTrackButton = await screen.findByRole('button', {name: 'no track'});
+    const primaryButton = await screen.findByRole('button', {name: 'primary'});
+    const secondaryButton = await screen.findByRole('button', {name: 'secondary'});
+    const dangerButton = await screen.findByRole('button', {name: 'danger'});
+    const buttonWithIcon = await screen.findByRole('button', {name: 'Take a photo'});
+    const noTrackLink = await screen.findByRole('link', {name: 'no track link'});
+    const link = await screen.findByRole('link', {name: 'link'});
+
+    userEvent.click(noTrackButton);
+    userEvent.click(noTrackLink);
+    expect(logEventSpy).not.toHaveBeenCalled();
+
+    userEvent.click(primaryButton);
+    expect(logEventSpy).toHaveBeenCalledWith({
+        category: 'user_interaction',
+        action: 'primary_button_tapped',
+        label: 'primary',
+    });
+    expect(logEventSpy).toHaveBeenCalledTimes(1);
+
+    userEvent.click(secondaryButton);
+    expect(logEventSpy).toHaveBeenCalledWith({
+        category: 'user_interaction',
+        action: 'secondary_button_tapped',
+        label: 'secondary',
+    });
+    expect(logEventSpy).toHaveBeenCalledTimes(2);
+
+    userEvent.click(dangerButton);
+    expect(logEventSpy).toHaveBeenCalledWith({
+        category: 'user_interaction',
+        action: 'danger_button_tapped',
+        label: 'danger',
+    });
+    expect(logEventSpy).toHaveBeenCalledTimes(3);
+
+    userEvent.click(buttonWithIcon);
+    expect(logEventSpy).toHaveBeenCalledWith({
+        category: 'user_interaction',
+        action: 'primary_button_tapped',
+        label: 'Take a photo',
+    });
+    expect(logEventSpy).toHaveBeenCalledTimes(4);
+
+    userEvent.click(link);
+    expect(logEventSpy).toHaveBeenCalledWith({
+        category: 'user_interaction',
+        action: 'link_tapped',
+        label: 'link',
+    });
+    expect(logEventSpy).toHaveBeenCalledTimes(5);
 });